+++ tests/test_component/test_attributes.py
@@ -15,4 +15,4 @@ def test_component_key_generation(tetra_request):
     """Test that the component key is generated correctly."""
 
     c = PersonComponent(tetra_request)
-    assert c.key == "main__lib__person_component"  # noqa
+    assert c.key.startswith("tk_")